uridownloader: use buffer lists to store fragment data
The maximum number of GstMemory blocks in a GstBuffer is 16, and
appending a 17th block triggers a merge of all GstMemory blocks to a
single block. For a large downloaded fragment this means a lot of
memcpys are going to happen.

To avoid copies, switch to GstBufferList to store downloaded buffers,
but still merge the buffers until the maximum number of memory blocks is
reached.

The gst_fragment_get_buffer function is still left for compatibility or
simplicity. It allocates a single buffer and copies the full content of
the buffer list in this new buffer.